+++ Commons/TransactionApi/src/main/java/com/mfino/transactionapi/handlers/account/impl/MFASubscriberActivationHandlerImpl.java
@@ -320,7 +320,7 @@ public class MFASubscriberActivationHandlerImpl extends FIXMessageHandler implem
 							+ subscriber.getId());
 				}
 				subscriber.setUpgradablekyclevel(null);
-                subscriber.setUpgradestate(CmFinoFIX.UpgradeState_Approved.longValue());
+                subscriber.setUpgradestate(CmFinoFIX.UpgradeState_Approved);
 				subscriber.setApproveorrejectcomment("Approved for No Emoney");
 				subscriber.setApprovedorrejectedby("System");
 				subscriber.setApproveorrejecttime(new Timestamp());
